Corrected errors related to selecting, adding, deleting, and editing ProductTemplateParts. Some screen placement modifications to the views. Added the Parts constraints (part number must start with a 'P', cannot be deleted from the database if associated with a product template).

+++ src/assignment2/PartsInventoryGateway.java
@@ -164,6 +164,9 @@ public class PartsInventoryGateway {
 			if (isPartAssociatedWithInventoryItem(partID)) {
 				throw new SQLException("Error: Cannot delete part associated with an inventory item."); // "Failed to delete entry..."
 			}
+			if (isPartAssociatedWithProductTemplate(partID)) {
+				throw new SQLException("Error: Cannot delete part associated with a product template."); // "Failed to delete entry..."
+			}
 			SQL = "DELETE FROM Parts WHERE ID=? ";
 			prepstmt = conn.prepareStatement(SQL);
 			prepstmt.setInt(1, partID);
@@ -320,6 +323,32 @@ public class PartsInventoryGateway {
 		createConnection();
 		SQL = "SELECT InventoryItems.ID FROM InventoryItems ";
 		SQL += "WHERE InventoryItems.PartID=?";
+		try {
+			prepstmt = conn.prepareStatement(SQL);
+			prepstmt.setInt(1, partID);
+			rs = prepstmt.executeQuery();
+			if (rs.next()) {
+				closeResultSet();
+				closePreparedStatement();
+				return true;
+			}
+			else {
+				closeResultSet();
+				closePreparedStatement();
+				return false;
+			}
+		} catch (SQLException e1) {
+			closeResultSet();
+			closePreparedStatement();
+			e1.printStackTrace();
+			return false;
+		}
+	}
+	
+	private boolean isPartAssociatedWithProductTemplate(Integer partID) throws SQLException, IOException {
+		createConnection();
+		SQL = "SELECT ProductTemplateParts.ID FROM ProductTemplateParts ";
+		SQL += "WHERE ProductTemplateParts.PartID=?";
 		try {
 			prepstmt = conn.prepareStatement(SQL);
 			prepstmt.setInt(1, partID);
